refactor doc endpoint security/status
