+++ app/api/security_stats.py
@@ -1,6 +1,7 @@
 """
-Security API endpoints.
+Security Statistics Endpoint
 """
+
 from datetime import datetime
 from fastapi import APIRouter
 
@@ -8,175 +9,6 @@ from ..core.logger import get_logger
 
 logger = get_logger(__name__)
 router = APIRouter()
-
-
-@router.get(
-    "/status",
-    tags=["Security"],
-    summary="🛡️ Security System Status",
-    description="""
-<h2>🔒 Advanced Security Protection Status</h2>
-
-<p>Provides real-time status of the multi-layer security system protecting the API
-from various attacks and abuse patterns.</p>
-
-<h3>🛡️ <strong>Protection Features Monitored</strong></h3>
-
-<ul>
-<li><strong>Multi-layer Rate Limiting</strong>: IP-based and global request limits</li>
-<li><strong>Attack Detection</strong>: Sophisticated pattern recognition</li>
-<li><strong>Request Fingerprinting</strong>: Unique request identification</li>
-<li><strong>Behavioral Analysis</strong>: User behavior pattern analysis</li>
-<li><strong>IP Switching Detection</strong>: Rapid IP change detection</li>
-<li><strong>Duplicate Prevention</strong>: File and request deduplication</li>
-<li><strong>Global Attack Scoring</strong>: Overall threat level assessment</li>
-</ul>
-
-<h3>📊 <strong>Status Information</strong></h3>
-
-<ul>
-<li><strong>System Status</strong>: Overall security system health</li>
-<li><strong>Active Protections</strong>: List of enabled security features</li>
-<li><strong>Threat Level</strong>: Current global attack probability</li>
-<li><strong>Storage Backend</strong>: In-memory or Redis storage status</li>
-<li><strong>Performance Metrics</strong>: Security system performance data</li>
-</ul>
-
-<h3>🎯 <strong>Use Cases</strong></h3>
-
-<ul>
-<li><strong>Security Monitoring</strong>: Real-time protection status</li>
-<li><strong>Threat Assessment</strong>: Current security posture</li>
-<li><strong>System Health</strong>: Security component validation</li>
-<li><strong>Compliance</strong>: Security audit and reporting</li>
-</ul>
-
-<h3>⚡ <strong>Performance</strong></h3>
-
-<ul>
-<li><strong>Response Time</strong>: &lt; 100ms typical</li>
-<li><strong>Rate Limiting</strong>: No limits applied to security endpoints</li>
-<li><strong>Real-time Data</strong>: Live security metrics</li>
-</ul>
-    """,
-    response_description="Current security system status and active protections",
-    responses={
-        200: {
-            "description": "Security status retrieved successfully",
-            "content": {
-                "application/json": {
-                    "example": {
-                        "service": "Pneumonia Detection API",
-                        "security_status": "active",
-                        "timestamp": "2025-09-05T10:30:00.000Z",
-                        "advanced_protection": {
-                            "global_attack_score": 0.15,
-                            "requests_per_minute": 23,
-                            "recent_unique_ips": 8,
-                            "blocked_fingerprints": 2,
-                            "storage_type": "memory"
-                        },
-                        "protection_features": [
-                            "Multi-layer Rate Limiting (In-Memory)",
-                            "IP Switching Attack Detection (In-Memory)",
-                            "Request Fingerprinting (In-Memory)",
-                            "Behavioral Analysis (In-Memory)",
-                            "Global Attack Scoring (In-Memory)",
-                            "Duplicate File Detection (In-Memory)"
-                        ]
-                    }
-                }
-            }
-        },
-        500: {
-            "description": "Security system error",
-            "content": {
-                "application/json": {
-                    "example": {
-                        "service": "Pneumonia Detection API",
-                        "security_status": "not_initialized",
-                        "timestamp": "2025-09-05T10:30:00.000Z",
-                        "error": "Rate limiter not initialized"
-                    }
-                }
-            }
-        }
-    }
-)
-async def get_security_status():
-    """
-    **🛡️ Advanced Security System Status**
-    
-    Provides comprehensive real-time status of the multi-layer security protection
-    system including threat detection, rate limiting, and attack prevention measures.
-    
-    **Security Layers Monitored:**
-    - **Rate Limiting**: Request frequency controls per IP
-    - **Attack Detection**: Sophisticated attack pattern recognition
-    - **Behavioral Analysis**: User behavior anomaly detection
-    - **Request Fingerprinting**: Unique request identification and tracking
-    - **IP Switching Detection**: Rapid IP change pattern detection
-    - **File Duplication Prevention**: Duplicate upload detection
-    - **Global Threat Scoring**: Overall attack probability assessment
-    
-    **Status Categories:**
-    - **active**: Security system fully operational
-    - **not_initialized**: Security components not ready
-    - **degraded**: Partial security functionality
-    
-    **Returns:**
-        dict: Complete security status with:
-        - Overall system health and operational status
-        - List of active protection features
-        - Current threat levels and attack scores
-        - Storage backend status and performance
-        - Timestamp for status validity
-        
-    **Use Cases:**
-        - Real-time security monitoring
-        - Threat level assessment
-        - Security audit and compliance
-        - System health verification
-    """
-    # Import at runtime to get the latest reference
-    from ..core.advanced_rate_limiting import get_rate_limiter
-    advanced_rate_limiter = get_rate_limiter()
-    
-    if advanced_rate_limiter is None:
-        return {
-            "service": "Pneumonia Detection API",
-            "security_status": "not_initialized",
-            "timestamp": datetime.now().isoformat(),
-            "error": "Rate limiter not initialized"
-        }
-    
-    # Use async method if available, otherwise fallback
-    try:
-        if hasattr(advanced_rate_limiter, '_storage_initialized') and advanced_rate_limiter._storage_initialized:
-            status = await advanced_rate_limiter.get_security_status_async()
-        else:
-            status = advanced_rate_limiter.get_security_status()
-    except Exception as e:
-        logger.error(f"Failed to get security status: {e}")
-        status = {"error": str(e)}
-    
-    return {
-        "service": "Pneumonia Detection API",
-        "security_status": "active",
-        "timestamp": datetime.now().isoformat(),
-        "advanced_protection": status,
-        "protection_features": [
-            "Multi-layer Rate Limiting (In-Memory)",
-            "IP Switching Attack Detection (In-Memory)", 
-            "Request Fingerprinting (In-Memory)",
-            "Behavioral Analysis (In-Memory)",
-            "Global Attack Scoring (In-Memory)",
-            "Duplicate File Detection (In-Memory)",
-            "Persistent Storage (In-Memory)",
-            "Single-instance Optimized (In-Memory)"
-        ]
-    }
-
 
 @router.get(
     "/stats",
